requirements.txt
asgiref==3.8.1
barcode==1.0.4
chardet==5.2.0
diff-match-patch==20241021
Django==5.1.5
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-filter==24.3
django-import-export==4.3.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
feedparser==6.0.11
gunicorn==23.0.0
Levenshtein==0.26.1
Markdown==3.7
numpy==2.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psycopg2==2.9.10
PyJWT==2.10.1
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-Levenshtein==0.26.1
python-slugify==8.0.4
pytz==2025.1
RapidFuzz==3.12.1
reportlab==4.3.0
setuptools==75.8.0
sgmllib3k==1.0.0
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
text-unidecode==1.3
tzdata==2025.1
